pyproject.toml
[tool.poetry]
name = "usgs"
version = "0.1.0"
description = ""
authors = ["Sayan Mitra snmitra41@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
dynaconf = "^3.1.11"
pandas = "^1.5.1"
psycopg2-binary = "^2.9.5"
pydantic = "^1.10.2"
jinja2 = "^3.1.2"
ipykernel = "^6.17.1"
loguru = "^0.6.0"
pytest = "^7.2.0"
black = "^22.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
